backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.4.0
python-docx==1.1.2
bcrypt==4.3.0
